chore: clean the cuda-feature clippy debt to zero
CI lints without the cuda feature, so the GPU tree had quietly
accumulated warnings that only a feature-enabled clippy run surfaces:
redundant casts, manual divisibility checks, collapsible lets, doc
lines that markdown mistook for lists, unused bench locals, and one
backward entry point past the argument budget - its per-layer negated-A
pointer now rides in the layer bundle where the other per-layer data
lives. cargo clippy --features cuda --all-targets -D warnings is clean;
the full battery and the hash ledgers are unchanged.

+++ src/mamba3_siso/gpu/inference.rs
@@ -1253,7 +1253,7 @@ impl Mamba3GpuInferenceMixed {
                 bld.arg(&ds_i);
                 let eps_g5: f32 = engine.cfg.rms_norm_eps;
                 bld.arg(&eps_g5);
-                let src_stride = (ng_i as i32) * ds_i;
+                let src_stride = ng_i * ds_i;
                 bld.arg(&src_stride);
                 unsafe { bld.launch(grid) }.map_err(|e| format!("M3 F4a bcnorm B+C: {e:?}"))?;
             }
